Trait key_store::traits::KeyStoreValue

pub trait KeyStoreValue {
    type Error;
    type SerializedValue;
    fn serialize(&self) -> Result<Self::SerializedValue, Self::Error>;

    fn deserialize(raw: &mut [u8]) -> Result<Self, Self::Error>
    where
        Self: Sized;
}

Any value that is stored in the key store must implement this trait. In most cases these are the raw bytes of the object.

Associated Types

type Error

The error type returned by the KeyStoreValue.

type SerializedValue

The type of a serialized key store value.

Loading content...
Required methods

fn serialize(&self) -> Result<Self::SerializedValue, Self::Error>

Serialize the value and return it as byte vector.

Returns an Error if the serialization fails.

fn deserialize(raw: &mut [u8]) -> Result<Self, Self::Error> where
Self: Sized,

Deserialize the byte slice and return the object.

Returns an Error if the deserialization fails.

Loading content...