Module key_store::types

Key Store Types

This module defines the Status enum. The Status defines values to tag key store values with.

XXX: We might want to make the Status a trait type as well 🤔.

Enums

Status
The status of a value in the key store.

StatusError
Errors thrown by operations on the Status.